+++ core/src/test/kotlin/net/corda/core/flows/TestDataVendingFlow.kt
@@ -2,7 +2,7 @@ package net.corda.core.flows
 
 import co.paralleluniverse.fibers.Suspendable
 import net.corda.core.identity.Party
-import net.corda.core.internal.FetchDataFlow
+import net.corda.core.node.flows.FetchDataFlow
 import net.corda.core.node.flows.SendStateAndRefFlow
 import net.corda.core.utilities.UntrustworthyData
