+++ docs/source/example-code/src/main/java/net/corda/docs/FlowCookbookJava.java
@@ -8,7 +8,7 @@ import net.corda.core.crypto.SecureHash;
 import net.corda.core.crypto.TransactionSignature;
 import net.corda.core.flows.*;
 import net.corda.core.identity.Party;
-import net.corda.core.internal.FetchDataFlow;
+import net.corda.core.node.flows.FetchDataFlow;
 import net.corda.core.node.flows.*;
 import net.corda.core.node.services.ServiceType;
 import net.corda.core.node.services.Vault;
