+++ node-api/src/test/kotlin/net/corda/nodeapi/internal/serialization/KryoTests.kt
@@ -7,7 +7,7 @@ import com.esotericsoftware.kryo.io.Output
 import com.google.common.primitives.Ints
 import net.corda.core.contracts.PrivacySalt
 import net.corda.core.crypto.*
-import net.corda.core.internal.FetchDataFlow
+import net.corda.core.node.flows.FetchDataFlow
 import net.corda.core.serialization.*
 import net.corda.core.utilities.ProgressTracker
 import net.corda.core.utilities.sequence
